✨ Unify horizontal and vertical ticks (#4)
The way I did it before, the horizontal and vertical ticks had different
cells and thus weren't equal in size. This fixes that by making
everything one large grid of squares. The reason I'm not using an actual
css grid is because the spaces between cells have a rounding error when
you use a lot of cells. Example: https://play.tailwindcss.com/xHAKyDzMfJ