📱 Make layout responsive
This is quite a significant layout change where we fully move towards a
grid with calculated cell sizes. This is important because otherwise the
grid calculated by default always gets odd pixels after the division and
then the cells aren't squares. We do this now from the inside out,
always guaranteeing that each cell is a square